wip: move approval to summary modal
