+++ src/hooks/useMaxAmountIn.ts
@@ -1,0 +1,13 @@
+import { Currency, CurrencyAmount, Percent, Token, TradeType } from '@uniswap/sdk-core'
+import { useMemo } from 'react'
+import { InterfaceTrade } from 'state/routing/types'
+
+export function useMaxAmountIn(
+  trade: InterfaceTrade<Currency, Currency, TradeType> | undefined,
+  allowedSlippage: Percent
+) {
+  return useMemo(() => {
+    const maximumAmountIn = trade?.maximumAmountIn(allowedSlippage)
+    return maximumAmountIn?.currency.isToken ? (maximumAmountIn as CurrencyAmount<Token>) : undefined
+  }, [allowedSlippage, trade])
+}
